Develop : example (#22)
* Split fypp file for `src`
The related .f90 file is not changed.

* Update fypp file for `src`
The associated .f90 file has been modified,
but the behavior of the implemented functions remain unchanged.

* Update fypp file for `test`
The associated .f90 files has been modified,
but the behavior of the implemented subroutines remain unchanged.

* Split fypp file for `src`
The associated .f90 file is not changed.

* Update `is_ieee_*`
Add `function`s for `ieee_class_type`

* Add an example files

* Update a manifest file for fpm
Update a step to run an example

* Update the example
Add the target `module`

* Update the example
Update the source file for `set_ieee_*`